Update to work with the latest version of libuv

+++ features/step_definitions/additional_cli_steps.rb
@@ -10,11 +10,13 @@ Then /^the output should contain consumed workload$/ do
 end
 
 Given /^a named pipe "(.*?)"$/ do |path|
-  if RUBY_PLATFORM.downcase.include?("mswin")
-    f = File.open(path, 'w+')
-    f.close
-  else
+  require 'ffi'
+  if not FFI::Platform.windows?
     system "/usr/bin/mkfifo", path
+    at_exit { File.unlink(path) }
   end
-  at_exit { File.unlink(path) }
+end
+
+Given /^I wait for (\d+) seconds?$/ do |n|
+  sleep(n.to_i)
 end
